Extract the benchmark test fixture
Move the local benchmark-mode Fedify target used by fedify bench integration
tests into test/bench.  This gives the scenario tests a shared fixture app
that exercises WebFinger, actor discovery, signature verification, and inbox
handling through the same benchmark target.

https://github.com/fedify-dev/fedify/issues/744
https://github.com/fedify-dev/fedify/issues/784
